rearrange localvolume code
