+++ src/localvolume/item.h
@@ -88,71 +88,4 @@ public:
     virtual std::unique_ptr<OpenFile> openRF() = 0;
 };
 
-class PlainFileItem : public FileItem
-{
-public:
-    using FileItem::FileItem;
-
-    virtual FInfo getFInfo()
-    {
-        return FInfo {
-            TICKX("TEXT"),
-            TICKX("ttxt"),
-            CWC(0),  // fdFlags
-            { CWC(0), CWC(0) },   // fdLocation
-            CWC(0)   // fdFldr
-        };
-    }
-
-    virtual void setFInfo(FInfo finfo)
-    {
-    }
-
-
-    virtual std::unique_ptr<OpenFile> open();
-    virtual std::unique_ptr<OpenFile> openRF();
-};
-
-class BasiliskFileItem : public FileItem
-{
-public:
-    using FileItem::FileItem;
-
-    virtual FInfo getFInfo();
-    virtual void setFInfo(FInfo finfo);
-    virtual std::unique_ptr<OpenFile> open();
-    virtual std::unique_ptr<OpenFile> openRF();
-
-    virtual void deleteItem();
-    virtual void renameItem(mac_string_view newName);
-    virtual void moveItem(const fs::path& newParent);
-};
-
-class AppleDoubleFileItem : public FileItem
-{
-public:
-    using FileItem::FileItem;
-
-    virtual FInfo getFInfo();
-    virtual void setFInfo(FInfo finfo);
-    virtual std::unique_ptr<OpenFile> open();
-    virtual std::unique_ptr<OpenFile> openRF();
-
-    virtual void deleteItem();
-    virtual void renameItem(mac_string_view newName);
-    virtual void moveItem(const fs::path& newParent);
-};
-
-class MacFileItem : public FileItem
-{
-public:
-    using FileItem::FileItem;
-
-    virtual FInfo getFInfo();
-    virtual void setFInfo(FInfo finfo);
-    virtual std::unique_ptr<OpenFile> open();
-    virtual std::unique_ptr<OpenFile> openRF();
-};
-
-
 }
