+++ src/localvolume/localvolume.h
@@ -157,28 +157,6 @@ public:
     virtual ItemPtr handleDirEntry(const DirectoryItem& parent, const fs::directory_entry& e);
 };
 
-class BasiliskHandler : public MetaDataHandler
-{
-    LocalVolume& volume;
-public:
-    BasiliskHandler(LocalVolume& vol) : volume(vol) {}
-    virtual bool isHidden(const fs::directory_entry& e);
-    virtual ItemPtr handleDirEntry(const DirectoryItem& parent, const fs::directory_entry& e);
-};
-
-class AppleDoubleHandler : public MetaDataHandler
-{
-    LocalVolume& volume;
-public:
-    AppleDoubleHandler(LocalVolume& vol) : volume(vol) {}
-    virtual bool isHidden(const fs::directory_entry& e);
-    virtual ItemPtr handleDirEntry(const DirectoryItem& parent, const fs::directory_entry& e);
-};
-
-class MacHandler : public MetaDataHandler
-{
-    virtual ItemPtr handleDirEntry(const DirectoryItem& parent, const fs::directory_entry& e);
-};
 
 class OpenFile
 {
@@ -191,48 +169,12 @@ public:
     virtual size_t write(size_t offset, void *p, size_t n) { throw OSErrorException(wrPermErr); }
 };
 
-class PlainDataFork : public OpenFile
-{
-    //fs::fstream stream;
-    int fd;
-    fs::path path_;
-public:
-    PlainDataFork(fs::path path);
-    ~PlainDataFork();
-
-    virtual size_t getEOF() override;
-    virtual void setEOF(size_t sz) override;
-    virtual size_t read(size_t offset, void *p, size_t n) override;
-    virtual size_t write(size_t offset, void *p, size_t n) override;
-};
 
 class EmptyFork : public OpenFile
 {
 public:
     virtual size_t getEOF() override { return 0; }
     virtual size_t read(size_t offset, void *p, size_t n) override { return 0; }
-};
-
-class AppleSingleDoubleFork : public OpenFile
-{
-    struct EntryDescriptor
-    {
-        GUEST_STRUCT;
-        GUEST<uint32_t> entryId;
-        GUEST<uint32_t> offset;
-        GUEST<uint32_t> length;
-    };
-    std::unique_ptr<OpenFile> file;
-    EntryDescriptor desc;
-    uint32_t descOffset;
-public:
-    AppleSingleDoubleFork(std::unique_ptr<OpenFile> file, int entry);
-    ~AppleSingleDoubleFork();
-
-    virtual size_t getEOF() override;
-    virtual void setEOF(size_t sz) override;
-    virtual size_t read(size_t offset, void *p, size_t n) override;
-    virtual size_t write(size_t offset, void *p, size_t n) override;
 };
