+++ src/localvolume/mac.cpp
@@ -1,6 +1,5 @@
+#include "mac.h"
 #include "host-os-config.h"
-#include "localvolume.h"
-#include "item.h"
 
 #ifdef MACOSX
 using namespace Executor;
