+++ src/localvolume/plain.cpp
@@ -1,5 +1,5 @@
+#include "plain.h"
 #include "host-os-config.h"
-#include "localvolume.h"
 #include <unistd.h>
 #include <sys/types.h>
 #include <sys/stat.h>
